Refactor campaigns and donations
