+++ backend/src/config/constants.ts
@@ -1,3 +1,1 @@
 export const CAMPAIGN_OPTIONS = ['DRAFT', 'PUBLISHED', 'ARCHIVE'];
-
-
