+++ backend/src/models/Donation.model.ts
@@ -1,7 +1,9 @@
 import mongoose from 'mongoose';
 import uniqueValidator from 'mongoose-unique-validator';
 
-const DonationSchema = new mongoose.Schema(
+import { IDonation } from '../interfaces/models/Donation';
+
+const DonationSchema = new mongoose.Schema<IDonation>(
   {
     transactionId: { type: String, required: true, unique: true },
     campaignId: {
@@ -36,7 +38,7 @@ const DonationSchema = new mongoose.Schema(
       },
     },
     isAnonymous: { type: Boolean, required: true, default: false },
-    emailReciept: { type: String },
+    emailReceipt: { type: String },
     isVerified: { type: Boolean, required: true, default: false },
     amount: { type: Number, required: true },
   },
@@ -58,4 +60,4 @@ DonationSchema.set('toJSON', {
 
 DonationSchema.plugin(uniqueValidator);
 
-export default mongoose.model('Donation', DonationSchema);
+export default mongoose.model<IDonation>('Donation', DonationSchema);
